Introducing Catch2 as a testframework for the project and a simple Expected type
The Expected implementation is rather simple. Neither an overload for expected
"void", nor for an expected reference - both overloads are not expected
to be missing for this project.
The tests for the expected are not exhaustive either. Even though
r-value overloads are supplied within the expected the implementation is
not tested.